[AESH-267] AeshCommandInvocation should receive the interface on constructor and not the implementation.
This reverts commit ffe250e2980ce32174750d79ba2db40d6e506287.
