+++ src/main/java/org/jboss/aesh/console/AeshConsole.java
@@ -89,4 +89,6 @@ public interface AeshConsole {
     boolean isRunning();
 
     ExportManager getExportManager();
+
+    Console getConsole();
 }
